Abstract world screen
