+++ core/src/net/mostlyoriginal/game/screen/OdbFeatureScreen.java
@@ -1,64 +1,24 @@
 package net.mostlyoriginal.game.screen;
 
 import com.artemis.World;
-import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.math.MathUtils;
+import net.mostlyoriginal.api.utils.builder.WorldBuilder;
+import net.mostlyoriginal.game.screen.core.WorldScreen;
 import net.mostlyoriginal.game.system.ClearScreenSystem;
 import net.mostlyoriginal.game.system.detection.OdbFeatureDetectionSystem;
 
 /**
- * Shows all detected artemis-odb features.
+ * Shows all detected artemis-odb features for a couple of seconds.
  *
  * @author Daan van Yperen
  */
-public class OdbFeatureScreen implements Screen {
+public class OdbFeatureScreen extends WorldScreen {
 
-	public static final float MIN_DELTA = 1 / 15f;
-
-	protected final World world;
-
-	public OdbFeatureScreen() {
-
-		world = new World();
-		world.setSystem(new ClearScreenSystem());
-		world.setSystem(new OdbFeatureDetectionSystem());
-		world.initialize();
-	}
-
-	@Override
-	public void show() {
-
-	}
-
-	@Override
-	public void render(float delta) {
-		// Prevent spikes in delta from causing insane world updates.
-		world.setDelta(MathUtils.clamp(delta, 0, MIN_DELTA));
-		world.process();
-	}
-
-	@Override
-	public void resize(int width, int height) {
-
-	}
-
-	@Override
-	public void pause() {
-
-	}
-
-	@Override
-	public void resume() {
-
-	}
-
-	@Override
-	public void hide() {
-
-	}
-
-	@Override
-	public void dispose() {
-
+	protected World createWorld() {
+		World world = new WorldBuilder()
+				.with(
+						new ClearScreenSystem(),
+						new OdbFeatureDetectionSystem()
+				).initialize();
+		return world;
 	}
 }
